These 3 core IUP files needed changes for Android.
iup_mask.c and iup_str.c use localeconv() and locale_info->decimal_point[0] which are not available on Android. (The common Lua trick which also has this problem is hard code '.' as the decimal point.

iup_object.h introduces struct _jobject as the InativeHandle type for Android. There is stuff commented out for a AndroidWrapper struct...I was originally thinking I might need more information stored than just the widget, but for the moment, just the Object is sufficient.