These are the Android Java files needed (so far) to complete the implementation for getting IupDialog going. Also includes extra C files containing Java->C (JNI) functions.
The IUP design is as follows:

We need a custom Application class (IupApplication) and a launcher Activity class (IupLaunchActivity).
The Application class might be unnecessary right now with the launcer Activity class, but it does provide us the possibility of monitoring Activity changes and low memory warnings.

The IupLaunchActivity will get created after the IupApplication. For now, I've decided to put the System.LoadLibrary stuff for native modules here, to follow the SDL convention. (I'm debating if this should go in the Application class. Another potential advantage of the Activity is that if the load takes awhile, the Activity might be able to show a splash screen...doing in the Application might be too early for this.) Also following SDL inspiration, I have a method called getLibraries which is intended to be a list of libraries users can list as strings with the intention that they can subclass this and change the list without modifying the original file.
We need this intermediate launch activity because Iup lets you start a program without creating an IupDialog, but Android must always have an Activity. In the manifest (and optionally with the_intent.setFlags(Intent.FLAG_ACTIVITY_NO_HISTORY), we specify that this Activity has no back stack history. So when a user creates a new IupDialog (Activity), that one will load and this launcher one will completely disappear from the user's perspective. This also makes this Activity a prime candiate for subclassing with the purpose of making it a splash screen.

IupActivity is the Activity that maps to what IupDialog creates. This is the one that is directly tied into the normal IupDialog functionality. There is a lot of hoop jumping because we have to go back and forth between C and Java. Android API's are also quite poor at letting us just access data directly. For example, when startActivity to create a new one, we don't actually get an object back. We must wait for the Android system to create it later and override the onCreate method for that class where we can finally get the "this" pointer to the object. Additionally, Activites can't communiticate directly to each other, and need to go through an Intent class. In this case, I use the Intent to pass the C IHandle through so I can set the ih->handle in the onCreate callback since I can't actually get the object when Iup tries to create it.

IupCommon is a static class providing some common utility helper functions.

Note: All Activities must be listed in the AndroidManifest.xml. This is an Android rule.

Both IupApplication and IupLaunchActivity classes are intended to be subclassed by end users to suit any specific needs without requiring them to muck with the IUP implementation.

WARNING: We cannot use IupMainLoop(). There is no way to support this because Android/Java completely controls the application loop and there is no way to take control of this. This is going to require us to come up with a different way of initializing IUP. I'm thinking something like the idle callback model where we must supply a callback function to get called to run our initialization code when the OS deems it ready. Similarly, we might need a close callback.

There is currently a hard coded function called IupEntry which will need to be changed.