Android: Added more groundwork code for providing a callback for application backgrounding. (Still need to connect to actual IUP callback system.)
Thinking out loud: This backgrounding event (for true) is probably the equivalent for Quit/Exit on mobile. Both Android and iOS allow you to swap out of the app, and there is no guarantee you will come back to it or formally exit it. In fact on iOS, there is no explicit exit. Developers are expected to save needed state on the background event.