Android: Implemented initial Button and lots of refinements for Activity.
Made a Java wrapper for iupAttribGet so I can use it in Java to set the TITLE properties on the Activity and Button.
Made changes to the manifest so screen rotation won't kill/restart the new Activities, and also so I can see the title bar on the Activity.
Make unMap for IupActivity call finish() because I think I may need to do this for the case where the user calls IupDestroy on the Dialog when it is still in the system. This is untested.
Made IupApplication accessible as a singleton so I can easily access a Application Context.

I had a problem with IUP dealing with the fact that Android does not give you back an Activity object immediately with startActivity.
When I create a new Android Activity, I do not get back an object
immediately. I have to wait for the system to create the Activity some
time later after this procedure returns. So in the Iup Dialog map
function callback, I can create a new activity, but I cannot yet set
the Ihandle->handle to anything yet.

Meanwhile, I expect the Button map function callback to be invoked,
but it appears that IUP's internals are skipping this because my
ih->handle for the parent is still NULL.

I have a workaround that probably violates every
principle Android tried to over-engineer. When I start a new activity, I
also create a root GroupView (which in my case is an instance of their
RelativeLayout manager which I intend to make interoperate with IUP's
layout system.
I originally was going to create this root view with the onCreate
callback method since I must attach it to the new Activity, but I
discovered I can create a detached GroupView and use it as a stand in
for Ihandle->handle. The button will now get Mapped and put into the
GroupView which gets me around that original problem and still seems
to work. Then the final trick is to figure out how to get the
GroupView attached to the Activity (and swap the handles around) at
the right time. So here is the really dirty trick. GroupView is not
serializable so I can't pass it through an Intent. However, since I am
using C and JNI, I can do things nobody expected me to do. I can pass
the C pointer address to the Ihandle (as a Java long) through the
Intent. In the Activity OnCreate callback, I can convert back the Java
long to the pointer address and go back through C/JNI to get me the
Ihandle->handle which has the saved Java reference to the GroupView. I
can then attach the view and swap the pointers around. Thus, I
effectively passed a non-serializable Java object between Activities.