Android: Button ACTION callback is now implemented. Common routine handleIupCallback implemented to handle all callbacks from Java. Test program now opens a new Dialog when the button is pushed.
Discovered that my idea for saving the Activity callstack isn't so hot because it's actually really hard to get the current activity from a view (e.g. button). So I now do the shadow variable thing in the IupApplication class so I can always get a current activity.

I use weak references so I don't accidentally retain it. However, it also means it could theoretically become NULL which will break things if called in that state. (I'm not sure if that is possible.)

TODO: Remove the callframe stuff.