Android: Implemented proposal for new IUP startup sequence because the existing one cannot work on Android and iOS.
On Android, there is no int main() function and we need an entry point into C from Java.
The new convention looks for a C function called IupEntryPoint. This can be changed by setting meta-data in the AndroidManifest:
        <meta-data android:name="ENTRY_POINT" android:value="IupEntryPoint" />

The code will look for this entry. If not found, it will fallback to IupEntryPoint(). The symbol is looked up by name using dlsym to avoid creating a linker dependency between the Iup library and the main library.

(The user may also specify a callback function by calling
IupSetFunction("ENTRY_POINT", (Icallback)IupEntryPoint);
This is how I expect other platforms to set/change the callback. But on Android, this code path is pretty impractical.)

The Android implementation also calls IupOpen() on behalf for the user since there isn't a good place for it.

So for a typical cross-platform project, I expect something like:

// TODO: Proposed exit callback which the user should setup in the IupEntryPoint() callback.
void IupExitCallback()
{
	IupClose();
}

// This is where user code should start
void IupEntryPoint()
{
	// TODO: Proposed exit callback so the user can call IupClose()
	IupSetFunction("EXIT_CB", (Icallback)IupExitCallback);

	Ihandle* button = IupButton("Iup Button", "");
	IupSetCallback(button, "ACTION", (Icallback)OnButtonCallback);
	Ihandle* dialog = IupDialog(button);
	IupShow(dialog);
}

// Remember that main is never run on Android.
// This should be a template/copy-and-paste for all IUP projects.
// Don't put any user code here. All user code should start in IupEntryPoint.
int main(int argc, char* argv[])
{
	IupOpen(0, NULL);
	IupSetFunction("ENTRY_POINT", (Icallback)IupEntryPoint);
	// User can no longer assume that IupMainLoop blocks, and may return immediately.
        // You should probably not put code after this call
	IupMainLoop();
	return 0;
}