Android: I think Android 7.0 intentionally broke RTLD_DEFAULT behavior as their way of trying to block access to private libs. We don't use private libs, but in some cases, I'm getting back NULL in some cases on Android 7.0 devices. I'm unclear why, because in isolated cases built directly into the Iup Android Studio project, this still works. But taking a built .aar and .so and using it in another project on 64-bit Android 7.0 devices seems to break. Maybe it could be that test builds are treated differently than product builds. So as a fallback/workaround, if the pointer is still NULL by this point, we will try to directly dlopen the library to get the handle.
Assumption: System.loadLibrary() has already loaded our library.
I presume since it is loaded by loadLibrary, dlopen() might not need to be as rigorous with absolute paths
or other possible side-effects which could break loadLibrary if we reversed the order.
(which would require apriori information to where these things are since there is no official API for this).

To support this change, there are new requirements for downstream users. Subclasses of IupLaunchApplication must @override getEntryPointLibraryName(). The default uses libMyIupProgram.so. Alternatively, there is a new Manifest setting called ENTRY_LIBRARY.