Android: Need to cache IupApplication class in JNI to deal with potential threading issues.
I am testing a mechanism to deal with threads and plan to propose for
IUP to formally adopt. The problem is when you are on a
non-main-ui-thread, and need to signal back to the main ui thread to
safely call UI APIs.

The overall idea is that we will provide an API that lets us
PostMessage(user_data) from any thread. This will send an event to the
UI thread and trigger it to run a callback on the main UI thread.

For Android, there are multiple challenges.
- We need a Context to get a Looper to create a Handler which can post
and run something on the UI thread.
- We need to use JNI to get both the Context and call into code to
create the Handler.
	- Which means we need to use AttachCurrentThread
	- Which hits this problem this commit fixes where FindClass()
fails if we created and Attached our own thread.

Anyway, this seems to fix the problem. IupApplication is special enough
(since it is used to get a Context) that it is probably worth caching as
a special case.