Android: Experimenting with new GetEnv variations so the user can know if their thread had to be Attached, so they can immediately Detach the thread if they choose.
Users with high performance threads may not like getting their thread attached because the JVM can potentially pause their threads, such as during garbage collection.
But we must attach their thread in order to do anything useful for native API calls.
So a compromise is to attach their thread, and then let them know if we had to attach it. Then they can choose to detach if they think it is worth it.
I have not measured the cost of attaching/detaching threads.